+++ src/index.ts
@@ -4,4 +4,5 @@ export * from './offchain';
 export * from './request';
 export * from './schema-encoder';
 export * from './schema-registry';
+export * from './transaction';
 export * from './utils';