+++ src/indexer.ts
@@ -2,10 +2,10 @@ import { Indexer__factory, Indexer as IndexerContract } from '@ethereum-attestat
 import { Overrides } from 'ethers';
 import { legacyVersion } from './legacy/version';
 import { DelegatedProxy } from './offchain';
-import { Base, SignerOrProvider, Transaction } from './transaction';
+import { Base, Transaction, TransactionSigner } from './transaction';
 
 export interface IndexerOptions {
-  signerOrProvider?: SignerOrProvider;
+  signer?: TransactionSigner;
 }
 
 export interface UIDOptions {
@@ -60,16 +60,16 @@ export class Indexer extends Base<IndexerContract> {
   private delegated?: DelegatedProxy;
 
   constructor(address: string, options?: IndexerOptions) {
-    const { signerOrProvider } = options || {};
+    const { signer } = options || {};
 
-    super(new Indexer__factory(), address, signerOrProvider);
+    super(new Indexer__factory(), address, signer);
   }
 
   // Connects the API to a specific signer
-  public connect(signerOrProvider: SignerOrProvider) {
+  public connect(signer: TransactionSigner) {
     delete this.delegated;
 
-    super.connect(signerOrProvider);
+    super.connect(signer);
 
     return this;
   }
@@ -86,9 +86,15 @@ export class Indexer extends Base<IndexerContract> {
 
   // Indexes an existing attestation
   public async indexAttestation({ uid }: IndexAttestationOptions, overrides?: Overrides): Promise<Transaction<void>> {
-    const tx = await this.contract.indexAttestation(uid, { ...overrides });
+    if (!this.signer) {
+      throw new Error('Invalid signer');
+    }
 
-    return new Transaction(tx, async () => {});
+    return new Transaction(
+      await this.contract.indexAttestation.populateTransaction(uid, { ...overrides }),
+      this.signer,
+      async () => {}
+    );
   }
 
   // Indexes multiple existing attestations
@@ -96,9 +102,15 @@ export class Indexer extends Base<IndexerContract> {
     { uids }: IndexAttestationsOptions,
     overrides?: Overrides
   ): Promise<Transaction<void>> {
-    const tx = await this.contract.indexAttestations(uids, { ...overrides });
+    if (!this.signer) {
+      throw new Error('Invalid signer');
+    }
 
-    return new Transaction(tx, async () => {});
+    return new Transaction(
+      await this.contract.indexAttestations.populateTransaction(uids, { ...overrides }),
+      this.signer,
+      async () => {}
+    );
   }
 
   public isAttestationIndexed({ uid }: IsAttestationIndexedOptions, overrides?: Overrides): Promise<boolean> {